downloader: Refactor GitRepo to correctly display its requirements
Previously, GitRepo did use the command name and version from Git, which
led to wrong output in the "requirements" command. Refactor GitRepo to
use its own command name, but do not implement getting the version at
all, as getting the "repo" tool version (not the version of the wrapper
to bootstrap the "repo" tool) is only possible after running "repo
init" to initalize a working tree.

The refactoring made GitBase so thin that it does not make sense to keep
it, so remove it for the sake of very little code duplication.
