--- downloader/src/main/kotlin/vcs/GitBase.kt
@@ -1,49 +1,0 @@
-/*
- * Copyright (C) 2017-2019 HERE Europe B.V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * License-Filename: LICENSE
- */
-
-package org.ossreviewtoolkit.downloader.vcs
-
-import org.ossreviewtoolkit.downloader.VersionControlSystem
-import org.ossreviewtoolkit.downloader.WorkingTree
-import org.ossreviewtoolkit.utils.CommandLineTool
-
-import java.io.File
-import java.util.regex.Pattern
-
-abstract class GitBase : VersionControlSystem(), CommandLineTool {
-    private val versionRegex = Pattern.compile("[Gg]it [Vv]ersion (?<version>[\\d.a-z-]+)(\\s.+)?")
-
-    override val defaultBranchName = "master"
-    override val latestRevisionNames = listOf("HEAD", "@")
-
-    override fun command(workingDir: File?) = "git"
-
-    override fun getVersion() = super.getVersion(null)
-
-    override fun transformVersion(output: String) =
-        versionRegex.matcher(output.lineSequence().first()).let {
-            if (it.matches()) {
-                it.group("version")
-            } else {
-                ""
-            }
-        }
-
-    override fun getWorkingTree(vcsDirectory: File): WorkingTree = GitWorkingTree(vcsDirectory, this)
-}
